+++ lib/screens/pages/stats_page.dart
@@ -1,9 +1,12 @@
+import 'package:auto_size_text/auto_size_text.dart';
 import 'package:clay_containers/clay_containers.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:infinity_sweeper/constants/style_constant.dart';
+import 'package:infinity_sweeper/helpers/gamedata_helper.dart';
+import 'package:infinity_sweeper/models/gamedata_model.dart';
 import 'package:infinity_sweeper/screens/components/app_bar.dart';
-import 'package:infinity_sweeper/screens/components/piechart.dart';
+import 'package:infinity_sweeper/screens/components/stats_pie_chart.dart';
 
 class StatsPage extends StatefulWidget {
   const StatsPage({Key? key}) : super(key: key);
@@ -13,9 +16,23 @@ class StatsPage extends StatefulWidget {
 }
 
 class _StatsPageState extends State<StatsPage> {
+  GameData gameData = GameData();
+
+  void initizialize() async {
+    gameData = await loadData();
+    setState(() {});
+  }
+
+  @override
+  void initState() {
+    super.initState();
+    initizialize();
+  }
+
   @override
   Widget build(BuildContext context) {
     SystemChrome.setEnabledSystemUIMode(SystemUiMode.immersive);
+
     Size size = MediaQuery.of(context).size;
     return Scaffold(
       appBar: TopBar(size, "Game statistics"),
@@ -23,8 +40,8 @@ class _StatsPageState extends State<StatsPage> {
       body: Center(
         child: Padding(
           padding: const EdgeInsets.only(
-            left: 8,
-            right: 8,
+            left: 20,
+            right: 20,
             top: 20,
             bottom: 20,
           ),
@@ -37,7 +54,48 @@ class _StatsPageState extends State<StatsPage> {
                 surfaceColor: StyleConstant.mainColor,
                 parentColor: StyleConstant.mainColor,
                 color: StyleConstant.mainColor,
-                child: const PieChartStats(),
+                child: PieChartStats(gameData),
+              ),
+              const SizedBox(height: 20),
+              ClayContainer(
+                borderRadius: 5,
+                curveType: CurveType.none,
+                surfaceColor: StyleConstant.mainColor,
+                parentColor: StyleConstant.mainColor,
+                color: StyleConstant.mainColor,
+                child: Padding(
+                  padding: const EdgeInsets.all(15.0),
+                  child: Column(
+                    children: [
+                      Row(
+                        mainAxisAlignment: MainAxisAlignment.start,
+                        children: const [
+                          AutoSizeText(
+                            "Your record is: ",
+                            style: TextStyle(
+                              fontWeight: FontWeight.bold,
+                              fontSize: 16,
+                              color: StyleConstant.textColor,
+                            ),
+                          ),
+                        ],
+                      ),
+                      Row(
+                        mainAxisAlignment: MainAxisAlignment.end,
+                        children: [
+                          AutoSizeText(
+                            formatHHMMSS(gameData.recordTimeInSecond),
+                            style: const TextStyle(
+                              fontWeight: FontWeight.bold,
+                              fontSize: 20,
+                              color: Colors.black,
+                            ),
+                          ),
+                        ],
+                      )
+                    ],
+                  ),
+                ),
               ),
               const SizedBox(height: 20),
               IconButton(
@@ -49,5 +107,24 @@ class _StatsPageState extends State<StatsPage> {
         ),
       ),
     );
+  }
+
+  String formatHHMMSS(int seconds) {
+    if (seconds != 0) {
+      int hours = (seconds / 3600).truncate();
+      seconds = (seconds % 3600).truncate();
+      int minutes = (seconds / 60).truncate();
+
+      String hoursStr = (hours).toString().padLeft(2, '0');
+      String minutesStr = (minutes).toString().padLeft(2, '0');
+      String secondsStr = (seconds % 60).toString().padLeft(2, '0');
+
+      if (hours == 0) {
+        return "$minutesStr:$secondsStr";
+      }
+      return "$hoursStr:$minutesStr:$secondsStr";
+    } else {
+      return "";
+    }
   }
 }
